Move linkifier types into browser
