+++ src/Types.d.ts
@@ -6,7 +6,7 @@
 import { ITerminalOptions as IPublicTerminalOptions, IDisposable, IMarker, ISelectionPosition } from 'xterm';
 import { ICharset, IAttributeData, CharData } from 'common/Types';
 import { IEvent, IEventEmitter } from 'common/EventEmitter';
-import { IColorSet } from 'browser/Types';
+import { IColorSet, ILinkifier, ILinkMatcherOptions } from 'browser/Types';
 import { IOptionsService } from 'common/services/Services';
 import { IBuffer, IBufferSet } from 'common/buffer/Types';
 import { IParams } from 'common/parser/Types';
@@ -14,9 +14,6 @@ import { IParams } from 'common/parser/Types';
 export type CustomKeyEventHandler = (event: KeyboardEvent) => boolean;
 
 export type LineData = CharData[];
-
-export type LinkMatcherHandler = (event: MouseEvent, uri: string) => void;
-export type LinkMatcherValidationCallback = (uri: string, callback: (isValid: boolean) => void) => void;
 
 /**
  * This interface encapsulates everything needed from the Terminal by the
@@ -167,27 +164,6 @@ export interface IInputHandler {
   /** ESC # 8 */ screenAlignmentPattern(): void;
 }
 
-export interface ILinkMatcher {
-  id: number;
-  regex: RegExp;
-  handler: LinkMatcherHandler;
-  hoverTooltipCallback?: LinkMatcherHandler;
-  hoverLeaveCallback?: () => void;
-  matchIndex?: number;
-  validationCallback?: LinkMatcherValidationCallback;
-  priority?: number;
-  willLinkActivate?: (event: MouseEvent, uri: string) => boolean;
-}
-
-export interface ILinkifierEvent {
-  x1: number;
-  y1: number;
-  x2: number;
-  y2: number;
-  cols: number;
-  fg: number;
-}
-
 export interface ITerminal extends IPublicTerminal, IElementAccessor, IBufferAccessor, ILinkifierAccessor {
   screenElement: HTMLElement;
   browser: IBrowser;
@@ -285,51 +261,6 @@ export interface ITerminalOptions extends IPublicTerminalOptions {
   useFlowControl?: boolean;
 }
 
-export interface ILinkifier {
-  onLinkHover: IEvent<ILinkifierEvent>;
-  onLinkLeave: IEvent<ILinkifierEvent>;
-  onLinkTooltip: IEvent<ILinkifierEvent>;
-
-  attachToDom(element: HTMLElement, mouseZoneManager: IMouseZoneManager): void;
-  linkifyRows(start: number, end: number): void;
-  registerLinkMatcher(regex: RegExp, handler: LinkMatcherHandler, options?: ILinkMatcherOptions): number;
-  deregisterLinkMatcher(matcherId: number): boolean;
-}
-
-export interface ILinkMatcherOptions {
-  /**
-   * The index of the link from the regex.match(text) call. This defaults to 0
-   * (for regular expressions without capture groups).
-   */
-  matchIndex?: number;
-  /**
-   * A callback that validates an individual link, returning true if valid and
-   * false if invalid.
-   */
-  validationCallback?: LinkMatcherValidationCallback;
-  /**
-   * A callback that fires when the mouse hovers over a link.
-   */
-  tooltipCallback?: LinkMatcherHandler;
-  /**
-   * A callback that fires when the mouse leaves a link that was hovered.
-   */
-  leaveCallback?: () => void;
-  /**
-   * The priority of the link matcher, this defines the order in which the link
-   * matcher is evaluated relative to others, from highest to lowest. The
-   * default value is 0.
-   */
-  priority?: number;
-  /**
-   * A callback that fires when the mousedown and click events occur that
-   * determines whether a link will be activated upon click. This enables
-   * only activating a link when a certain modifier is held down, if not the
-   * mouse event will continue propagation (eg. double click to select word).
-   */
-  willLinkActivate?: (event: MouseEvent, uri: string) => boolean;
-}
-
 export interface IBrowser {
   isNode: boolean;
   userAgent: string;
@@ -339,21 +270,4 @@ export interface IBrowser {
   isIpad: boolean;
   isIphone: boolean;
   isWindows: boolean;
-}
-
-export interface IMouseZoneManager extends IDisposable {
-  add(zone: IMouseZone): void;
-  clearAll(start?: number, end?: number): void;
-}
-
-export interface IMouseZone {
-  x1: number;
-  x2: number;
-  y1: number;
-  y2: number;
-  clickCallback: (e: MouseEvent) => any;
-  hoverCallback: (e: MouseEvent) => any | undefined;
-  tooltipCallback: (e: MouseEvent) => any | undefined;
-  leaveCallback: () => any | undefined;
-  willLinkActivate: (e: MouseEvent) => boolean;
 }
