+++ src/renderer/dom/DomRenderer.ts
@@ -4,11 +4,11 @@
  */
 
 import { IRenderer, IRenderDimensions, CharacterJoinerHandler } from 'browser/renderer/Types';
-import { ILinkifierEvent, ITerminal } from '../../Types';
+import { ITerminal } from '../../Types';
 import { BOLD_CLASS, ITALIC_CLASS, CURSOR_CLASS, CURSOR_STYLE_BLOCK_CLASS, CURSOR_BLINK_CLASS, CURSOR_STYLE_BAR_CLASS, CURSOR_STYLE_UNDERLINE_CLASS, DomRendererRowFactory } from 'browser/renderer/dom/DomRendererRowFactory';
 import { INVERTED_DEFAULT_COLOR } from 'browser/renderer/atlas/Constants';
 import { Disposable } from 'common/Lifecycle';
-import { IColorSet } from 'browser/Types';
+import { IColorSet, ILinkifierEvent } from 'browser/Types';
 import { ICharSizeService } from 'browser/services/Services';
 import { IOptionsService } from 'common/services/Services';
